uni_chain_api_simu/requirements.txt
asn1crypto==0.24.0
base58==1.0.3
cffi==1.12.3
Django==2.2.1
django-rest-framework==0.1.0
djangorestframework==3.9.3
djangorestframework-jwt==1.11.0
Pillow==6.0.0
psycopg2-binary==2.8.2
pycparser==2.19
PyJWT==1.7.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
